+++ sdk/storage/Azure.Storage.Files.DataLake/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added support for service version 2021-06-08.
 - Added support for Customer Provided Key server-side encryption of files.
 - Added ability to retrieve path CreatedOn and ExpiresOn times with DataLakeFileSystemClient.GetPaths() and .GetPathsAsync().
+- Added support for DataLakeFileSystemClient.GetPathClient().
 
 ## 12.9.0 (2022-03-10)
 - Includes all features from 12.9.0-beta.1, 12.9.0-beta.2, and 12.9.0-beta.3 except SDK-calculated transactional checksums on data transfer.